pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pot"
version = "0.2.0"
description = "Pointer-over-Heads: Dynamic Routing Transformer for Structured NLP"
readme = "README.md"
authors = [
    {name = "Eran Ben Artzy", email = "eran@example.com"}
]
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["transformers", "attention", "routing", "nlp", "structured-prediction"]

dependencies = [
    "torch>=1.13.0",
    "transformers>=4.25.0",
    "datasets>=2.8.0",
    "pyyaml>=6.0",
    "numpy>=1.20.0",
    "scipy>=1.7.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "isort>=5.10.0",
]

[project.urls]
Homepage = "https://github.com/Eran-BA/PoT"
Repository = "https://github.com/Eran-BA/PoT"
Documentation = "https://github.com/Eran-BA/PoT/tree/main/docs"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src/pot --cov-report=term-missing"
